requirements.txt
libnum==1.7.1
pycryptodome==3.21.0
